--- src/main/java/umc/study/spring/apiPayLoad/code/BaseCode.java
@@ -1,6 +1,0 @@
-package umc.study.spring.apiPayLoad.code;
-
-public interface BaseCode {
-    public ReasonDTO getReason();
-    public ReasonDTO getReasonHttpStatus();
-}
