--- src/main/java/umc/study/spring/apiPayLoad/exception/handler/CategoryHandler.java
@@ -1,10 +1,0 @@
-package umc.study.spring.apiPayLoad.exception.handler;
-
-import umc.study.spring.apiPayLoad.code.BaseErrorCode;
-import umc.study.spring.apiPayLoad.exception.GeneralException;
-
-public class CategoryHandler extends GeneralException {
-    public CategoryHandler(BaseErrorCode code) {
-        super(code);
-    }
-}
