--- src/main/java/umc/study/spring/apiPayLoad/exception/handler/StoreHandler.java
@@ -1,10 +1,0 @@
-package umc.study.spring.apiPayLoad.exception.handler;
-
-import umc.study.spring.apiPayLoad.code.BaseErrorCode;
-import umc.study.spring.apiPayLoad.exception.GeneralException;
-
-public class StoreHandler extends GeneralException {
-    public StoreHandler(BaseErrorCode code) {
-        super(code);
-    }
-}
