+++ src/main/java/umc/study/spring/apiPayload/code/BaseCode.java
@@ -1,0 +1,6 @@
+package umc.study.spring.apiPayload.code;
+
+public interface BaseCode {
+    public umc.study.spring.apiPayload.code.ReasonDTO getReason();
+    public umc.study.spring.apiPayload.code.ReasonDTO getReasonHttpStatus();
+}
