+++ src/main/java/umc/study/spring/apiPayload/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package umc.study.spring.apiPayLoad.code;
+package umc.study.spring.apiPayload.code;
 
 public interface BaseErrorCode {
     public ErrorReasonDTO getReason();
