+++ src/main/java/umc/study/spring/apiPayload/code/status/SuccessStatus.java
@@ -1,10 +1,10 @@
-package umc.study.spring.apiPayLoad.code.status;
+package umc.study.spring.apiPayload.code.status;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import umc.study.spring.apiPayLoad.code.BaseCode;
-import umc.study.spring.apiPayLoad.code.ReasonDTO;
+import umc.study.spring.apiPayload.code.ReasonDTO;
+import umc.study.spring.apiPayload.code.BaseCode;
 
 @Getter
 @AllArgsConstructor
