+++ src/main/java/umc/study/spring/apiPayload/exception/handler/CategoryHandler.java
@@ -1,0 +1,10 @@
+package umc.study.spring.apiPayload.exception.handler;
+
+import umc.study.spring.apiPayload.code.BaseErrorCode;
+import umc.study.spring.apiPayload.exception.GeneralException;
+
+public class CategoryHandler extends GeneralException {
+    public CategoryHandler(BaseErrorCode code) {
+        super(code);
+    }
+}
