+++ src/main/java/umc/study/spring/apiPayload/exception/handler/FoodCategoryHandler.java
@@ -1,0 +1,10 @@
+package umc.study.spring.apiPayload.exception.handler;
+
+import umc.study.spring.apiPayload.code.BaseErrorCode;
+import umc.study.spring.apiPayload.exception.GeneralException;
+
+public class FoodCategoryHandler extends GeneralException {
+    public FoodCategoryHandler(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+}
