+++ src/main/java/umc/study/spring/domain/mapping/UserMission.java
@@ -29,4 +29,8 @@ public class UserMission extends BaseEntity {
 
     @Column(nullable = false)
     private Integer checkCode;
+
+    public void setComplete() {
+        this.status = MissionStatus.COMPLETED;
+    }
 }
