+++ src/main/java/umc/study/spring/service/StoreService/StoreServiceImpl.java
@@ -3,8 +3,8 @@ package umc.study.spring.service.StoreService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.study.spring.apiPayLoad.code.status.ErrorStatus;
-import umc.study.spring.apiPayLoad.exception.handler.StoreHandler;
+import umc.study.spring.apiPayload.code.status.ErrorStatus;
+import umc.study.spring.apiPayload.exception.handler.StoreHandler;
 import umc.study.spring.domain.Store;
 import umc.study.spring.repository.StoreRepository;
