+++ src/main/java/umc/study/spring/service/UserService/UserCommandServiceImpl.java
@@ -3,8 +3,8 @@ package umc.study.spring.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.study.spring.apiPayLoad.code.status.ErrorStatus;
-import umc.study.spring.apiPayLoad.exception.handler.CategoryHandler;
+import umc.study.spring.apiPayload.code.status.ErrorStatus;
+import umc.study.spring.apiPayload.exception.handler.CategoryHandler;
 import umc.study.spring.converter.UserConverter;
 import umc.study.spring.converter.UserPreferConverter;
 import umc.study.spring.domain.Category;
