+++ src/main/java/umc/study/spring/validation/validator/CategoriesExistValidator.java
@@ -2,8 +2,7 @@ package umc.study.spring.validation.validator;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import umc.study.spring.apiPayLoad.code.status.ErrorStatus;
-import umc.study.spring.repository.CategoryRepository;
+import umc.study.spring.apiPayload.code.status.ErrorStatus;
 import umc.study.spring.service.CategoryService.CategoryService;
 import umc.study.spring.validation.annotation.ExistCategories;
