+++ src/main/java/umc/study/spring/web/controller/MissionRestController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import umc.study.spring.apiPayLoad.ApiResponse;
+import umc.study.spring.apiPayload.ApiResponse;
 import umc.study.spring.converter.MissionConverter;
 import umc.study.spring.domain.Mission;
 import umc.study.spring.service.MissionService.MissionCommandService;
