+++ src/main/java/umc/study/spring/web/controller/ReviewRestController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import umc.study.spring.apiPayLoad.ApiResponse;
+import umc.study.spring.apiPayload.ApiResponse;
 import umc.study.spring.converter.ReviewConverter;
 import umc.study.spring.domain.Review;
 import umc.study.spring.service.ReviewService.ReviewCommandService;
