+++ src/main/java/umc/study/spring/web/controller/TempRestController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import umc.study.spring.apiPayLoad.ApiResponse;
+import umc.study.spring.apiPayload.ApiResponse;
 import umc.study.spring.converter.TempConverter;
 import umc.study.spring.web.dto.TempResponse;
 import umc.study.spring.service.TempService.TempQueryService;
